api/requirements.txt
fastapi
uvicorn
pydantic
pypdf
google-generativeai
spacy
https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.4.0/pt_core_news_sm-3.4.0-py3-none-any.whl
